move cavc files to folders
