+++ client/app/queue/cavc/AddCavcRemandView.jsx
@@ -4,25 +4,25 @@ import { connect } from 'react-redux';
 import { css } from 'glamor';
 import _ from 'lodash';
 import PropTypes from 'prop-types';
-import COPY from '../../COPY';
-import CAVC_JUDGE_FULL_NAMES from '../../constants/CAVC_JUDGE_FULL_NAMES';
-import CAVC_REMAND_SUBTYPES from '../../constants/CAVC_REMAND_SUBTYPES';
-import CAVC_REMAND_SUBTYPE_NAMES from '../../constants/CAVC_REMAND_SUBTYPE_NAMES';
-import CAVC_DECISION_TYPES from '../../constants/CAVC_DECISION_TYPES';
+import COPY from '../../../COPY';
+import CAVC_JUDGE_FULL_NAMES from '../../../constants/CAVC_JUDGE_FULL_NAMES';
+import CAVC_REMAND_SUBTYPES from '../../../constants/CAVC_REMAND_SUBTYPES';
+import CAVC_REMAND_SUBTYPE_NAMES from '../../../constants/CAVC_REMAND_SUBTYPE_NAMES';
+import CAVC_DECISION_TYPES from '../../../constants/CAVC_DECISION_TYPES';
 
-import QueueFlowPage from './components/QueueFlowPage';
-import { requestSave, showErrorMessage } from './uiReducer/uiActions';
-import { validateDateNotInFuture } from '../intake/util/issues';
-import TextField from '../components/TextField';
-import RadioField from '../components/RadioField';
-import DateSelector from '../components/DateSelector';
-import Checkbox from '../components/Checkbox';
-import CheckboxGroup from '../components/CheckboxGroup';
-import TextareaField from '../components/TextareaField';
-import Button from '../components/Button';
-import SearchableDropdown from '../components/SearchableDropdown';
-import StringUtil from '../util/StringUtil';
-import Alert from '../components/Alert';
+import QueueFlowPage from '../components/QueueFlowPage';
+import { requestSave, showErrorMessage } from '../uiReducer/uiActions';
+import { validateDateNotInFuture } from '../../intake/util/issues';
+import TextField from '../../components/TextField';
+import RadioField from '../../components/RadioField';
+import DateSelector from '../../components/DateSelector';
+import Checkbox from '../../components/Checkbox';
+import CheckboxGroup from '../../components/CheckboxGroup';
+import TextareaField from '../../components/TextareaField';
+import Button from '../../components/Button';
+import SearchableDropdown from '../../components/SearchableDropdown';
+import StringUtil from '../../util/StringUtil';
+import Alert from '../../components/Alert';
 import { withRouter } from 'react-router';
 
 const radioLabelStyling = css({ marginTop: '2.5rem' });
